+++ scripts/src/machine.lua
@@ -2860,6 +2860,18 @@ end
 
 ---------------------------------------------------
 --
+--@src/devices/machine/lpc2132_vic.h,MACHINES["LPC2132_VIC"] = true
+---------------------------------------------------
+
+if MACHINES["LPC2132_VIC"] then
+	files {
+		MAME_DIR .. "src/devices/machine/lpc2132_vic.cpp",
+		MAME_DIR .. "src/devices/machine/lpc2132_vic.h",
+	}
+end
+
+---------------------------------------------------
+--
 --@src/devices/machine/lpci.h,MACHINES["LPCI"] = true
 ---------------------------------------------------
 
@@ -5580,18 +5592,3 @@ if MACHINES["QUADMOUSE"] then
 		MAME_DIR .. "src/devices/machine/quadmouse.h",
 	}
 end
-
----------------------------------------------------
-
---
---@src/devices/machine/lpc2132_vic.h,MACHINES["LPC2132_VIC"] = true
----------------------------------------------------
-
-if (MACHINES["LPC2132_VIC"]~=null) then
-	files {
-		MAME_DIR .. "src/devices/machine/lpc2132_vic.cpp",
-		MAME_DIR .. "src/devices/machine/lpc2132_vic.h",
-	}
-end
-
-
